requirements.txt
asgiref==3.6.0
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
constantly==15.1.0
cryptography==40.0.1
daphne==4.0.0
Django==4.1.7
django-cors-headers==3.14.0
django-multiselectfield==0.1.12
django-multiupload==0.6.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.1.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.7
zope.interface==6.0
